Refactor M3 reporter to support multiple encodings
